much better ergonomics for ingestContext
