+++ cmd/api/src/services/graphify/ingest_integration_test.go
@@ -23,354 +23,16 @@ import (
 	"bytes"
 	"errors"
 	"testing"
-	"time"
 
-	"github.com/specterops/bloodhound/cmd/api/src/daemons/changelog"
 	"github.com/specterops/bloodhound/cmd/api/src/model"
 	"github.com/specterops/bloodhound/cmd/api/src/services/graphify"
 	"github.com/specterops/bloodhound/cmd/api/src/services/upload"
 	"github.com/specterops/bloodhound/cmd/api/src/test/integration"
-	"github.com/specterops/bloodhound/packages/go/ein"
 	"github.com/specterops/bloodhound/packages/go/graphschema"
-	"github.com/specterops/bloodhound/packages/go/graphschema/ad"
 	"github.com/specterops/dawgs/graph"
 	"github.com/specterops/dawgs/query"
 	"github.com/stretchr/testify/require"
 )
-
-// verify that nodes and edges are created or updated based on existing graph state
-func Test_IngestRelationships(t *testing.T) {
-	t.Run("Create rel by exact name match. Source and target node names both resolve to nodes with objectids.", func(t *testing.T) {
-		testContext := integration.NewGraphTestContext(t, graphschema.DefaultGraphSchema())
-		bloodhoundDB := integration.OpenDatabase(t)
-
-		testContext.DatabaseTestWithSetup(
-			func(harness *integration.HarnessDetails) error {
-				harness.IngestRelationships.Setup(testContext)
-				return nil
-			},
-			func(harness integration.HarnessDetails, db graph.Database) {
-				ingestibleRel := ein.NewIngestibleRelationship(
-					ein.IngestibleEndpoint{Value: "computer a", Kind: ad.Computer, MatchBy: ein.MatchByName},
-					ein.IngestibleEndpoint{Value: "computer b", Kind: ad.Computer, MatchBy: ein.MatchByName},
-					ein.IngestibleRel{RelType: graph.StringKind("related_to")},
-				)
-				rels := []ein.IngestibleRelationship{ingestibleRel}
-
-				err := db.BatchOperation(testContext.Context(), func(batch graph.Batch) error {
-					cl := changelog.NewChangelog(db, bloodhoundDB, changelog.DefaultOptions())
-					ingestContext := graphify.NewIngestContext(testContext.Context(), batch, time.Now().UTC(), cl, false)
-
-					err := graphify.IngestRelationships(ingestContext, graph.EmptyKind, rels)
-					require.Nil(t, err)
-					return nil
-				})
-
-				require.Nil(t, err)
-
-				// verify an edge was created
-				err = db.ReadTransaction(testContext.Context(), func(tx graph.Transaction) error {
-					count, err := tx.Relationships().Filter(
-						query.And(
-							query.Equals(query.StartID(), harness.IngestRelationships.Node1.ID),
-							query.Equals(query.EndID(), harness.IngestRelationships.Node2.ID),
-							query.Kind(query.Relationship(), graph.StringKind("related_to")),
-						),
-					).Count()
-
-					require.Equal(t, int64(1), count)
-					require.Nil(t, err)
-
-					return nil
-				})
-
-				require.Nil(t, err)
-
-			})
-	})
-
-	t.Run("Update rel by exact name match. Source and target node names both resolve to nodes with objectids.", func(t *testing.T) {
-		testContext := integration.NewGraphTestContext(t, graphschema.DefaultGraphSchema())
-		bloodhoundDB := integration.OpenDatabase(t)
-
-		testContext.DatabaseTestWithSetup(
-			func(harness *integration.HarnessDetails) error {
-				harness.IngestRelationships.Setup(testContext)
-				return nil
-			},
-			func(harness integration.HarnessDetails, db graph.Database) {
-				ingestibleRel := ein.NewIngestibleRelationship(
-					ein.IngestibleEndpoint{Value: "computer a", Kind: ad.Computer, MatchBy: ein.MatchByName},
-					ein.IngestibleEndpoint{Value: "computer b", Kind: ad.Computer, MatchBy: ein.MatchByName},
-					ein.IngestibleRel{RelType: graph.StringKind("existing_edge_kind"), RelProps: map[string]any{"hello": "world"}},
-				)
-				rels := []ein.IngestibleRelationship{ingestibleRel}
-
-				err := db.BatchOperation(testContext.Context(), func(batch graph.Batch) error {
-					cl := changelog.NewChangelog(db, bloodhoundDB, changelog.DefaultOptions())
-					ingestContext := graphify.NewIngestContext(testContext.Context(), batch, time.Now().UTC(), cl, false)
-
-					err := graphify.IngestRelationships(ingestContext, graph.EmptyKind, rels)
-					require.Nil(t, err)
-					return nil
-				})
-
-				require.Nil(t, err)
-
-				// verify an edge was updated
-				err = db.ReadTransaction(testContext.Context(), func(tx graph.Transaction) error {
-					rels := []*graph.Relationship{}
-					err := tx.Relationships().Filter(
-						query.And(
-							query.Equals(query.StartID(), harness.IngestRelationships.Node1.ID),
-							query.Equals(query.EndID(), harness.IngestRelationships.Node2.ID),
-							query.Kind(query.Relationship(), graph.StringKind("existing_edge_kind")),
-						),
-					).Fetch(func(cursor graph.Cursor[*graph.Relationship]) error {
-						for rel := range cursor.Chan() {
-							rels = append(rels, rel)
-						}
-
-						return nil
-					})
-					require.Nil(t, err)
-					// was there only one exact match?
-					require.Len(t, rels, 1)
-					// were properties merged?
-					property, _ := rels[0].Properties.Get("hello").String()
-					require.Equal(t, "world", property)
-
-					return nil
-				})
-
-				require.Nil(t, err)
-
-			})
-	})
-
-	t.Run("Create rel using source/target nodes that specify objectid.", func(t *testing.T) {
-		testContext := integration.NewGraphTestContext(t, graphschema.DefaultGraphSchema())
-		bloodhoundDB := integration.OpenDatabase(t)
-
-		testContext.DatabaseTestWithSetup(
-			func(harness *integration.HarnessDetails) error {
-				harness.IngestRelationships.Setup(testContext)
-				return nil
-			},
-			func(harness integration.HarnessDetails, db graph.Database) {
-				ingestibleRel := ein.NewIngestibleRelationship(
-					ein.IngestibleEndpoint{Value: "1234"},
-					ein.IngestibleEndpoint{Value: "5678"},
-					ein.IngestibleRel{RelType: graph.StringKind("related_to")},
-				)
-				rels := []ein.IngestibleRelationship{ingestibleRel}
-
-				err := db.BatchOperation(testContext.Context(), func(batch graph.Batch) error {
-					cl := changelog.NewChangelog(db, bloodhoundDB, changelog.DefaultOptions())
-					ingestContext := graphify.NewIngestContext(testContext.Context(), batch, time.Now().UTC(), cl, false)
-
-					err := graphify.IngestRelationships(ingestContext, graph.EmptyKind, rels)
-					require.Nil(t, err)
-					return nil
-				})
-
-				require.Nil(t, err)
-
-				// verify an edge was created
-				err = db.ReadTransaction(testContext.Context(), func(tx graph.Transaction) error {
-					count, err := tx.Relationships().Filter(
-						query.And(
-							query.Equals(query.StartID(), harness.IngestRelationships.Node1.ID),
-							query.Equals(query.EndID(), harness.IngestRelationships.Node2.ID),
-							query.Kind(query.Relationship(), graph.StringKind("related_to")),
-						),
-					).Count()
-
-					require.Equal(t, int64(1), count)
-					require.Nil(t, err)
-
-					return nil
-				})
-
-				require.Nil(t, err)
-
-			})
-	})
-
-	t.Run("Update rel using source/target nodes that specify objectid.", func(t *testing.T) {
-		testContext := integration.NewGraphTestContext(t, graphschema.DefaultGraphSchema())
-		bloodhoundDB := integration.OpenDatabase(t)
-
-		testContext.DatabaseTestWithSetup(
-			func(harness *integration.HarnessDetails) error {
-				harness.IngestRelationships.Setup(testContext)
-				return nil
-			},
-			func(harness integration.HarnessDetails, db graph.Database) {
-				ingestibleRel := ein.NewIngestibleRelationship(
-					ein.IngestibleEndpoint{Value: "1234"},
-					ein.IngestibleEndpoint{Value: "5678"},
-					ein.IngestibleRel{RelType: graph.StringKind("existing_edge_kind"), RelProps: map[string]any{"hello": "world"}},
-				)
-				rels := []ein.IngestibleRelationship{ingestibleRel}
-
-				err := db.BatchOperation(testContext.Context(), func(batch graph.Batch) error {
-					cl := changelog.NewChangelog(db, bloodhoundDB, changelog.DefaultOptions())
-					ingestContext := graphify.NewIngestContext(testContext.Context(), batch, time.Now().UTC(), cl, false)
-
-					err := graphify.IngestRelationships(ingestContext, graph.EmptyKind, rels)
-					require.Nil(t, err)
-					return nil
-				})
-
-				require.Nil(t, err)
-
-				// verify an edge was updated
-				err = db.ReadTransaction(testContext.Context(), func(tx graph.Transaction) error {
-					rels := []*graph.Relationship{}
-					err := tx.Relationships().Filter(
-						query.And(
-							query.Equals(query.StartID(), harness.IngestRelationships.Node1.ID),
-							query.Equals(query.EndID(), harness.IngestRelationships.Node2.ID),
-							query.Kind(query.Relationship(), graph.StringKind("existing_edge_kind")),
-						),
-					).Fetch(func(cursor graph.Cursor[*graph.Relationship]) error {
-						for rel := range cursor.Chan() {
-							rels = append(rels, rel)
-						}
-
-						return nil
-					})
-					require.Nil(t, err)
-					// was there only one exact match?
-					require.Len(t, rels, 1)
-					// were properties merged?
-					property, _ := rels[0].Properties.Get("hello").String()
-					require.Equal(t, "world", property)
-
-					return nil
-				})
-
-				require.Nil(t, err)
-
-			})
-	})
-
-	t.Run("Create rel. Source/target nodes' objectid's dont exist. Both nodes get created and rel gets created.", func(t *testing.T) {
-		testContext := integration.NewGraphTestContext(t, graphschema.DefaultGraphSchema())
-		bloodhoundDB := integration.OpenDatabase(t)
-
-		testContext.DatabaseTestWithSetup(
-			func(harness *integration.HarnessDetails) error {
-				harness.IngestRelationships.Setup(testContext)
-				return nil
-			},
-			func(harness integration.HarnessDetails, db graph.Database) {
-				ingestibleRel := ein.NewIngestibleRelationship(
-					ein.IngestibleEndpoint{Value: "0001"},
-					ein.IngestibleEndpoint{Value: "0002"},
-					ein.IngestibleRel{RelType: graph.StringKind("related_to")},
-				)
-				rels := []ein.IngestibleRelationship{ingestibleRel}
-
-				err := db.BatchOperation(testContext.Context(), func(batch graph.Batch) error {
-					cl := changelog.NewChangelog(db, bloodhoundDB, changelog.DefaultOptions())
-					ingestContext := graphify.NewIngestContext(testContext.Context(), batch, time.Now().UTC(), cl, false)
-
-					err := graphify.IngestRelationships(ingestContext, graph.EmptyKind, rels)
-					require.Nil(t, err)
-					return nil
-				})
-
-				require.Nil(t, err)
-
-				err = db.ReadTransaction(testContext.Context(), func(tx graph.Transaction) error {
-					nodeIDs := map[string]graph.ID{}
-					// verify start and end nodes were created
-					_ = tx.Nodes().Filter(
-						query.Or(
-							query.Equals(query.Property(query.Node(), "objectid"), "0001"),
-							query.Equals(query.Property(query.Node(), "objectid"), "0002"),
-						)).
-						OrderBy(query.Order(query.Property(query.Node(), "objectid"), query.Ascending())).Fetch(
-						func(cursor graph.Cursor[*graph.Node]) error {
-							for node := range cursor.Chan() {
-								objectid, _ := node.Properties.Get("objectid").String()
-								nodeIDs[objectid] = node.ID
-							}
-							return nil
-						},
-					)
-
-					require.Len(t, nodeIDs, 2)
-
-					// verify rel created
-					count, err := tx.Relationships().Filter(
-						query.And(
-							query.Equals(query.StartID(), nodeIDs["0001"]),
-							query.Equals(query.EndID(), nodeIDs["0002"]),
-							query.Kind(query.Relationship(), graph.StringKind("related_to")),
-						),
-					).Count()
-
-					require.Equal(t, int64(1), count)
-					require.Nil(t, err)
-
-					return nil
-				})
-
-				require.Nil(t, err)
-
-			})
-	})
-
-	t.Run("Dont create rel. Source/target nodes' have names that don't resolve to objectids. Neither node gets created and rel creation is skipped.", func(t *testing.T) {
-		testContext := integration.NewGraphTestContext(t, graphschema.DefaultGraphSchema())
-		bloodhoundDB := integration.OpenDatabase(t)
-
-		testContext.DatabaseTestWithSetup(
-			func(harness *integration.HarnessDetails) error {
-				harness.IngestRelationships.Setup(testContext)
-				return nil
-			},
-			func(harness integration.HarnessDetails, db graph.Database) {
-				ingestibleRel := ein.NewIngestibleRelationship(
-					ein.IngestibleEndpoint{Value: "bubba", MatchBy: ein.MatchByName},
-					ein.IngestibleEndpoint{Value: "lubba", MatchBy: ein.MatchByName},
-					ein.IngestibleRel{RelType: graph.StringKind("related_to")},
-				)
-				rels := []ein.IngestibleRelationship{ingestibleRel}
-
-				err := db.BatchOperation(testContext.Context(), func(batch graph.Batch) error {
-					cl := changelog.NewChangelog(db, bloodhoundDB, changelog.DefaultOptions())
-					ingestContext := graphify.NewIngestContext(testContext.Context(), batch, time.Now().UTC(), cl, false)
-
-					err := graphify.IngestRelationships(ingestContext, graph.EmptyKind, rels)
-					require.ErrorContains(t, err, "skipping invalid relationship")
-					return nil
-				})
-
-				require.Nil(t, err)
-
-				err = db.ReadTransaction(testContext.Context(), func(tx graph.Transaction) error {
-					// verify start and end nodes were not created
-					numNodes, _ := tx.Nodes().Filter(query.Or(
-						query.Equals(query.Property(query.Node(), "name"), "bubba"),
-						query.Equals(query.Property(query.Node(), "name"), "lubba"))).Count()
-					require.Zero(t, numNodes)
-
-					// verify rel not created
-					numRels, _ := tx.Relationships().Filter(query.Kind(query.Relationship(), graph.StringKind("related_to"))).Count()
-					require.Zero(t, numRels)
-
-					return nil
-				})
-
-				require.Nil(t, err)
-
-			})
-	})
-}
 
 // verifies that files that bypassed validation controls due to being uploaded as zips receive validation attention in the datapipe,
 // and that invalid files are not ingested into the graph
@@ -390,12 +52,9 @@ func Test_ReadFileForIngest(t *testing.T) {
 
 	t.Run("happy path. a file uploaded as a zip passes validation and is written to the graph", func(t *testing.T) {
 		testContext := integration.NewGraphTestContext(t, graphschema.DefaultGraphSchema())
-		bloodhoundDB := integration.OpenDatabase(t)
 
 		testContext.BatchTest(func(harness integration.HarnessDetails, batch graph.Batch) {
-
-			cl := changelog.NewChangelog(testContext.Graph.Database, bloodhoundDB, changelog.DefaultOptions())
-			ingestContext := graphify.NewIngestContext(testContext.Context(), batch, time.Now().UTC(), cl, false)
+			ingestContext := graphify.NewIngestContext(testContext.Context(), batch)
 
 			err := graphify.ReadFileForIngest(ingestContext, validReader, readOptions)
 			require.Nil(t, err)
@@ -431,12 +90,10 @@ func Test_ReadFileForIngest(t *testing.T) {
 
 	t.Run("failure path. a file uploaded as a zip fails validation and nothing is written to the graph", func(t *testing.T) {
 		testContext := integration.NewGraphTestContext(t, graphschema.DefaultGraphSchema())
-		bloodhoundDB := integration.OpenDatabase(t)
 
 		testContext.DatabaseTestWithSetup(func(harness *integration.HarnessDetails) error { return nil }, func(harness integration.HarnessDetails, db graph.Database) {
 			_ = db.BatchOperation(testContext.Context(), func(batch graph.Batch) error {
-				cl := changelog.NewChangelog(db, bloodhoundDB, changelog.DefaultOptions())
-				ingestContext := graphify.NewIngestContext(testContext.Context(), batch, time.Now().UTC(), cl, false)
+				ingestContext := graphify.NewIngestContext(testContext.Context(), batch)
 
 				err := graphify.ReadFileForIngest(ingestContext, invalidReader, readOptions)
 				require.NotNil(t, err)
@@ -459,54 +116,3 @@ func Test_ReadFileForIngest(t *testing.T) {
 		})
 	})
 }
-
-// 1. changelog disabled , ingest a node. created
-// 2. changelog disabled , ingest a node. update.
-// invalid case, 0 kinds
-// 4 . changelog enabled. creating a node still works
-// 5. changelog enabled. dedupe works.
-// 6. changelog enabled. update works.
-
-// new setup helper that does exactly what IngestNode needs
-// func TestIngestNode(t *testing.T) {
-// 	t.Parallel()
-
-// 	var (
-// 		ctx         = context.Background()
-// 		fixturePath = path.Join("fixtures", "IngestNode")
-// 		testSuite   = setupIntegrationTestSuite(t, fixturePath)
-// 		// todo: put the cl on the test suite struct
-// 		cl = changelog.NewChangelog(testSuite.GraphDB, testSuite.BHDatabase, changelog.DefaultOptions())
-// 	)
-
-// 	defer teardownIntegrationTestSuite(t, &testSuite)
-
-// 	t.Run("ingested node gets created", func(t *testing.T) {
-// 		// open up a batchOp
-// 		testSuite.GraphDB.BatchOperation(ctx,
-// 			func(batch graph.Batch) error {
-// 				ingestCtx := graphify.NewIngestContext(testSuite.Context, batch, time.Now(), cl, false)
-
-// 				// function under test
-// 				err := graphify.IngestNode(ingestCtx, graph.StringKind("hellobase"), ein.IngestibleNode{
-// 					ObjectID: "ABC123",
-// 					Labels:   []graph.Kind{graph.StringKind("labelA"), graph.StringKind("labelB")},
-// 					PropertyMap: map[string]any{
-// 						"hello": "world",
-// 						"1":     2,
-// 					},
-// 				})
-
-// 				require.NoError(t, err)
-
-// 				return nil
-// 			})
-
-// 		testSuite.GraphDB.ReadTransaction(ctx, func(tx graph.Transaction) error {
-// 			tx.Nodes().Fetch(func(cursor graph.Cursor[*graph.Node]) error {
-
-// 			})
-// 		})
-// 	})
-
-// }
